requirements.txt
streamlit
edge-tts
requests
python-dotenv
speechrecognition
pyaudio
pyautogui
pywhatkit
playsound==1.2.2